requirements.txt
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
idna==2.8
nose==1.3.7
numpy==1.18.1
py==1.8.1
requests==2.22.0
retry==0.9.2
selenium==3.141.0
six==1.14.0
spec==1.4.1
urllib3==1.25.8
opencv-python==4.1.0.25
facebook-wda==0.3.6
Appium-Python-Client==0.47
Pillow==7.0.0